requirements.txt
fastapi
uvicorn
gunicorn
supabase
python-dotenv
notion-client
python-jose[cryptography]
requests
